tests/e2e/pyproject.toml
[project]
name = "framecast-e2e-tests"
version = "0.0.1.dev0"
description = "End-to-end tests for Framecast API"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
    "faker>=21.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyJWT>=2.8.0",          # JWT token generation for E2E auth
    "asyncpg>=0.29.0",       # Direct database access for test seeding
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--asyncio-mode=auto",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "auth: authentication related tests",
    "teams: team management tests",
    "security: security validation tests",
    "invitation: invitation workflow tests",
    "error_handling: error handling and edge case tests",
]

[tool.coverage.run]
source = ["tests"]
omit = ["tests/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["faker"]
ignore_missing_imports = true
